# docs/contracts/openapi-backend.yaml
openapi: 3.0.3
info:
  title: 또박또박 Backend API
  version: 0.1.0
  description: |
    데이터·판정(규칙 엔진)·창구 요약서 저장을 담당한다. Spring Boot 3 + MyBatis + MySQL.
    - JSON 필드는 모두 snake_case (Jackson SNAKE_CASE 전역 설정).
    - 시각은 ISO-8601 (예: 2026-09-08T15:12:00).
    - 이 파일이 프론트·AI 서버와의 유일한 계약이다. 변경은 고현준 확인 후 팀 공지.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/summaries:
    post:
      tags: [summaries]
      summary: 창구 요약서 생성. 코드·번호표·준비물은 서버가 계산한다.
      description: |
        준비물 규칙: REQUEST 에 TRANSFER 가 있으면 신분증 필수, 통장/카드 권장.
        번호표는 목업(브랜치별 카운터). 코드는 6자리 숫자 유니크.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SummaryCreateRequest" }
            example: { $ref: "./examples/summary-create-request.json" }
      responses:
        "201":
          description: 생성된 요약서
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Summary" }
              example: { $ref: "./examples/summary.json" }
  /api/summaries/{summary_id}:
    get:
      tags: [summaries]
      summary: 요약서 상세
      parameters:
        - { name: summary_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: 요약서, content: { application/json: { schema: { $ref: "#/components/schemas/Summary" } } } }
        "404": { description: 없음 }
  /api/summaries/by-code/{code}:
    get:
      tags: [summaries]
      summary: 6자리 코드로 요약서 조회 (직원 화면 진입)
      parameters:
        - { name: code, in: path, required: true, schema: { type: string, pattern: "^[0-9]{6}$" } }
      responses:
        "200": { description: 요약서, content: { application/json: { schema: { $ref: "#/components/schemas/Summary" } } } }
        "404": { description: 없음 }
  /api/summaries/{summary_id}/status:
    get:
      tags: [summaries]
      summary: 처리 상태 (어르신 앱이 폴링해서 "처리됐어요" 음성 안내)
      parameters:
        - { name: summary_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SummaryStatus" }
...
components:
  schemas:
...
    SummaryCreateRequest:
      type: object
      required: [user_id]
      properties:
        user_id: { type: integer }
        session_id: { type: string, nullable: true }
        branch_name: { type: string, nullable: true, description: 없으면 users.home_branch }
        requests:
          type: array
          items: { $ref: "#/components/schemas/RequestItemPayload" }
        questions:
          type: array
          items: { $ref: "#/components/schemas/QuestionItemPayload" }
    RequestItemPayload:
      type: object
      required: [type, recipient_name, amount]
      properties:
        type: { type: string, enum: [TRANSFER] }
        recipient_counterparty_id: { type: integer, nullable: true }
        recipient_name: { type: string, example: 김철수 }
        recipient_relation: { type: string, nullable: true, example: 아들 }
        recipient_bank: { type: string, nullable: true }
        recipient_account_masked: { type: string, nullable: true, description: 서버가 counterparty 로 채운다 }
        amount: { type: integer, example: 200000 }
        note: { type: string, nullable: true }
        confirmed_by_user: { type: boolean, description: CONFIRM 단계에서 예 응답을 받았는지. false 면 400. }
    QuestionItemPayload:
      type: object
      required: [text]
      properties:
        transaction_id: { type: integer, nullable: true }
        text: { type: string, example: "'대한정보통신' 19,000원 출금이 무엇인지" }
...
    SummaryItem:
      type: object
      required: [id, section, ordinal, payload, handled]
      properties:
        id: { type: integer }
        section: { type: string, enum: [REQUEST, QUESTION, PREP] }
        ordinal: { type: integer }
        payload:
          description: section 별로 RequestItemPayload / QuestionItemPayload / PrepItemPayload
          type: object
        handled: { type: boolean }
        handled_at: { type: string, format: date-time, nullable: true }
    Summary:
      type: object
      required: [id, code, ticket_no, branch_name, status, user_name, created_at, items]
      properties:
        id: { type: integer }
        code: { type: string, example: "482913" }
        ticket_no: { type: integer, example: 12 }
        branch_name: { type: string, example: KB국민은행 종로지점 }
        status: { type: string, enum: [OPEN, IN_PROGRESS, DONE] }
        user_name: { type: string }
        user_id: { type: integer }
        created_at: { type: string, format: date-time }
        items: { type: array, items: { $ref: "#/components/schemas/SummaryItem" } }
    SummaryStatus:
      type: object
      required: [id, status, handled_count, total_count]
      properties:
        id: { type: integer }
        status: { type: string, enum: [OPEN, IN_PROGRESS, DONE] }
        handled_count: { type: integer }
        total_count: { type: integer, description: PREP 제외 }
        last_handled_at: { type: string, format: date-time, nullable: true }
...